// src/models/Multa.ts
export default interface Imulta{
    id: number;
    codigo: string;
    nombreInfractor: string;
    apellidoInfractor: string;
    cedulaInfractor: string;
    placaVehiculo: string;
    monto: number;
    fotoevidencia: string;
    tipoDeLicencia: string;
    motivo: string;
    latitud: string;
    longitud: string;
    timestamp: string;
}